move function and try and except
